cli/remote_add: Check URI is reachable before allowing it to be added
Use curl's handy "no body" option to effectively ping the URI to see
if it is reachable.

Previously if a URI wasn't reachable it would be allowed to be added
anyway and moss would emit a scary 'unable to mmap file' backtrace
when it attempted to call refreshremotes() on it.

Long term, we want moss-fetcher to handle this use case. Especially,
for existing genuine remotes.